Split configuration for Apollo so that tracing can go in the tracing module and metrics can go in the metrics module.
